+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 TBD: Version 0.7 (Jul ??, 2017)
 --------------------------------------------------
-- New DependencyParserRegistry. Change NeedyRunner to use it.
-- GrailsBuildConfigDependencyParser: Change includePlugins default to true.
+- New `DependencyParserRegistry`. Change `NeedyRunner` to use it.
+- `GrailsBuildConfigDependencyParser`: Change *includePlugins* default to true.
 - Move `DoNothing` into standalone top-level class file.
 - Include sample HTML report in distribution and reference from README.
+- New WildcardUtil.
 
 
 Version 0.6 (Jul 19, 2017)